move functions out of /app-api.ts
