+++ src/lib/db-helper.ts
@@ -1,5 +1,5 @@
 import { storageBucket } from "@/lib/db"
-import { DoneeInfo } from "@/types/db"
+import { DoneeInfo, User } from "@/types/db"
 
 export async function getImageAsDataUrl(url: string) {
   const file = await storageBucket.file(url).download()
@@ -22,5 +22,64 @@ export async function downloadImagesForDonee(
     ...donee,
     signature: signatureDataUrl,
     smallLogo: smallLogoDataUrl,
+  }
+}
+
+export function receiptReady(user: User): user is User & {
+  items: Required<User>["items"]
+  donee: Required<DoneeInfo>
+  date: Required<User>["date"]
+} {
+  const { items, donee, date } = user
+  const {
+    companyAddress,
+    companyName,
+    country,
+    registrationNumber,
+    signatoryName,
+    signature,
+    smallLogo,
+  } = donee || {}
+
+  return Boolean(
+    items &&
+      date &&
+      companyAddress &&
+      companyName &&
+      country &&
+      registrationNumber &&
+      signatoryName &&
+      signature &&
+      smallLogo
+  )
+}
+export function alreadyFilledIn(user: User | undefined): {
+  items: boolean
+  doneeDetails: boolean
+} {
+  if (!user) return { items: false, doneeDetails: false }
+
+  const { items, donee, date } = user
+  const {
+    companyAddress,
+    companyName,
+    country,
+    registrationNumber,
+    signatoryName,
+    signature,
+    smallLogo,
+  } = donee || {}
+
+  return {
+    items: Boolean(items && date),
+    doneeDetails: Boolean(
+      companyAddress &&
+        companyName &&
+        country &&
+        registrationNumber &&
+        signatoryName &&
+        signature &&
+        smallLogo
+    ),
   }
 }
