+++ src/pages/api/email.ts
@@ -6,12 +6,12 @@ import { renderToBuffer } from "@react-pdf/renderer"
 import { z } from "zod"
 
 import {
-  assertSessionIsQboConnected,
   AuthorisedHandler,
   createAuthorisedHandler,
   parseRequestBody,
-  receiptReady,
-} from "@/lib/app-api"
+} from "@/lib/util/request-server"
+import { receiptReady } from "@/lib/db-helper"
+import { assertSessionIsQboConnected } from "@/lib/util/next-auth-helper"
 import { WithBody, ReceiptPdfDocument } from "@/components/receipt"
 import { user } from "@/lib/db"
 import { Donation, getDonations } from "@/lib/qbo-api"
