+++ src/pages/api/stripe/update-subscription.ts
@@ -2,7 +2,11 @@ import { z } from "zod"
 import { ApiError } from "next/dist/server/api-utils"
 
 import { manageSubscriptionStatusChange, stripe } from "@/lib/stripe"
-import { AuthorisedHandler, createAuthorisedHandler, parseRequestBody } from "@/lib/app-api"
+import {
+  AuthorisedHandler,
+  createAuthorisedHandler,
+  parseRequestBody,
+} from "@/lib/util/request-server"
 import { user } from "@/lib/db"
 
 export const parser = z.object({
