+++ src/pages/details.tsx
@@ -6,7 +6,7 @@ import { Session, getServerSession } from "next-auth"
 import { authOptions } from "./api/auth/[...nextauth]"
 import { buttonStyling } from "@/components/ui"
 import { user } from "@/lib/db"
-import { alreadyFilledIn } from "@/lib/app-api"
+import { alreadyFilledIn } from "@/lib/db-helper"
 import { base64EncodeFile, postJsonData } from "@/lib/util/request"
 import { DataType as DetailsApiDataType } from "@/pages/api/details"
 import { DoneeInfo } from "@/types/db"
@@ -134,7 +134,6 @@ export const getServerSideProps: GetServerSideProps<Props> = async context => {
   const dbUser = doc.data()
   if (!dbUser) throw new Error("User has no corresponding db entry")
 
-  // if donee data is already in db use that to prefill form otherwise use data from quickbooks
   return {
     props: {
       session,
