+++ src/pages/email.tsx
@@ -8,7 +8,8 @@ import { postJsonData } from "@/lib/util/request"
 import { MissingData } from "@/components/ui"
 import { Fieldset, TextArea } from "@/components/form"
 import { user } from "@/lib/db"
-import { alreadyFilledIn, isSessionQboConnected, receiptReady } from "@/lib/app-api"
+import { alreadyFilledIn, receiptReady } from "@/lib/db-helper"
+import { isSessionQboConnected } from "@/lib/util/next-auth-helper"
 import { DoneeInfo } from "@/types/db"
 import { getDonations } from "@/lib/qbo-api"
 import { isUserSubscribed } from "@/lib/stripe"
